pyproject.toml
[project]
name = "vulnerabilities"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "loguru>=0.7.3",
    "pyyaml>=6.0.2",
]
